Update Karma configuration to include WebkitHeadless
