--- engines/query-sparql/test/util-browser.js
@@ -1,5 +1,0 @@
-const fetchFn = globalThis.fetch;
-
-export function fetch(...args) {
-    return fetchFn(...args);
-}
