+++ engines/query-sparql/test/util-browser.ts
@@ -1,0 +1,5 @@
+const fetchFn = globalThis.fetch;
+
+export async function fetch(...args: Parameters<typeof fetchFn>): ReturnType<typeof fetchFn> {
+  return fetchFn(...args);
+}
